Create ExtensionManager class which manages all high level extension tasks
