+++ packages/cli/src/commands/extensions/disable.ts
@@ -5,11 +5,11 @@
  */
 
 import { type CommandModule } from 'yargs';
-import { disableExtension } from '../../config/extension.js';
-import { SettingScope } from '../../config/settings.js';
+import { loadSettings, SettingScope } from '../../config/settings.js';
 import { getErrorMessage } from '../../utils/errors.js';
-import { ExtensionEnablementManager } from '../../config/extensions/extensionEnablement.js';
 import { debugLogger } from '@google/gemini-cli-core';
+import { ExtensionManager } from '../../config/extension-manager.js';
+import { requestConsentNonInteractive } from '../../config/extensions/consent.js';
 
 interface DisableArgs {
   name: string;
@@ -17,20 +17,19 @@ interface DisableArgs {
 }
 
 export function handleDisable(args: DisableArgs) {
-  const extensionEnablementManager = new ExtensionEnablementManager();
+  const workspaceDir = process.cwd();
+  const extensionManager = new ExtensionManager({
+    workspaceDir,
+    enabledExtensionOverrides: [],
+    requestConsent: requestConsentNonInteractive,
+    loadedSettings: loadSettings(workspaceDir),
+  });
+
   try {
     if (args.scope?.toLowerCase() === 'workspace') {
-      disableExtension(
-        args.name,
-        SettingScope.Workspace,
-        extensionEnablementManager,
-      );
+      extensionManager.disableExtension(args.name, SettingScope.Workspace);
     } else {
-      disableExtension(
-        args.name,
-        SettingScope.User,
-        extensionEnablementManager,
-      );
+      extensionManager.disableExtension(args.name, SettingScope.User);
     }
     debugLogger.log(
       `Extension "${args.name}" successfully disabled for scope "${args.scope}".`,
