+++ packages/cli/src/commands/extensions/link.ts
@@ -6,15 +6,14 @@
 
 import type { CommandModule } from 'yargs';
 import {
-  installOrUpdateExtension,
-  requestConsentNonInteractive,
-} from '../../config/extension.js';
-import {
   debugLogger,
   type ExtensionInstallMetadata,
 } from '@google/gemini-cli-core';
 
 import { getErrorMessage } from '../../utils/errors.js';
+import { requestConsentNonInteractive } from '../../config/extensions/consent.js';
+import { ExtensionManager } from '../../config/extension-manager.js';
+import { loadSettings } from '../../config/settings.js';
 
 interface InstallArgs {
   path: string;
@@ -26,10 +25,15 @@ export async function handleLink(args: InstallArgs) {
       source: args.path,
       type: 'link',
     };
-    const extensionName = await installOrUpdateExtension(
-      installMetadata,
-      requestConsentNonInteractive,
-    );
+    const workspaceDir = process.cwd();
+    const extensionManager = new ExtensionManager({
+      workspaceDir,
+      enabledExtensionOverrides: [],
+      requestConsent: requestConsentNonInteractive,
+      loadedSettings: loadSettings(workspaceDir),
+    });
+    const extensionName =
+      await extensionManager.installOrUpdateExtension(installMetadata);
     debugLogger.log(
       `Extension "${extensionName}" linked successfully and enabled.`,
     );
