+++ packages/cli/src/commands/extensions/update.ts
@@ -6,10 +6,6 @@
 
 import type { CommandModule } from 'yargs';
 import {
-  loadExtensions,
-  requestConsentNonInteractive,
-} from '../../config/extension.js';
-import {
   updateAllUpdatableExtensions,
   type ExtensionUpdateInfo,
   checkForAllExtensionUpdates,
@@ -18,8 +14,10 @@ import {
 import { checkForExtensionUpdate } from '../../config/extensions/github.js';
 import { getErrorMessage } from '../../utils/errors.js';
 import { ExtensionUpdateState } from '../../ui/state/extensions.js';
-import { ExtensionEnablementManager } from '../../config/extensions/extensionEnablement.js';
 import { debugLogger } from '@google/gemini-cli-core';
+import { ExtensionManager } from '../../config/extension-manager.js';
+import { requestConsentNonInteractive } from '../../config/extensions/consent.js';
+import { loadSettings } from '../../config/settings.js';
 
 interface UpdateArgs {
   name?: string;
@@ -30,13 +28,15 @@ const updateOutput = (info: ExtensionUpdateInfo) =>
   `Extension "${info.name}" successfully updated: ${info.originalVersion} → ${info.updatedVersion}.`;
 
 export async function handleUpdate(args: UpdateArgs) {
-  const workingDir = process.cwd();
-  const extensionEnablementManager = new ExtensionEnablementManager(
-    // Force enable named extensions, otherwise we will only update the enabled
-    // ones.
-    args.name ? [args.name] : [],
-  );
-  const extensions = loadExtensions(extensionEnablementManager);
+  const workspaceDir = process.cwd();
+  const extensionManager = new ExtensionManager({
+    workspaceDir,
+    enabledExtensionOverrides: [],
+    requestConsent: requestConsentNonInteractive,
+    loadedSettings: loadSettings(workspaceDir),
+  });
+
+  const extensions = extensionManager.loadExtensions();
   if (args.name) {
     try {
       const extension = extensions.find(
@@ -54,7 +54,7 @@ export async function handleUpdate(args: UpdateArgs) {
       }
       const updateState = await checkForExtensionUpdate(
         extension,
-        extensionEnablementManager,
+        extensionManager,
       );
       if (updateState !== ExtensionUpdateState.UPDATE_AVAILABLE) {
         debugLogger.log(`Extension "${args.name}" is already up to date.`);
@@ -63,9 +63,7 @@ export async function handleUpdate(args: UpdateArgs) {
       // TODO(chrstnb): we should list extensions if the requested extension is not installed.
       const updatedExtensionInfo = (await updateExtension(
         extension,
-        extensionEnablementManager,
-        workingDir,
-        requestConsentNonInteractive,
+        extensionManager,
         updateState,
         () => {},
       ))!;
@@ -88,7 +86,7 @@ export async function handleUpdate(args: UpdateArgs) {
       const extensionState = new Map();
       await checkForAllExtensionUpdates(
         extensions,
-        extensionEnablementManager,
+        extensionManager,
         (action) => {
           if (action.type === 'SET_STATE') {
             extensionState.set(action.payload.name, {
@@ -96,14 +94,11 @@ export async function handleUpdate(args: UpdateArgs) {
             });
           }
         },
-        workingDir,
       );
       let updateInfos = await updateAllUpdatableExtensions(
-        workingDir,
-        requestConsentNonInteractive,
         extensions,
         extensionState,
-        extensionEnablementManager,
+        extensionManager,
         () => {},
       );
       updateInfos = updateInfos.filter(
