+++ packages/cli/src/commands/mcp/list.test.ts
@@ -16,19 +16,20 @@ import {
 } from 'vitest';
 import { listMcpServers } from './list.js';
 import { loadSettings } from '../../config/settings.js';
-import { ExtensionStorage, loadExtensions } from '../../config/extension.js';
 import { createTransport } from '@google/gemini-cli-core';
 import { Client } from '@modelcontextprotocol/sdk/client/index.js';
+import { ExtensionStorage } from '../../config/extensions/storage.js';
+import { ExtensionManager } from '../../config/extension-manager.js';
 
 vi.mock('../../config/settings.js', () => ({
   loadSettings: vi.fn(),
 }));
-vi.mock('../../config/extension.js', () => ({
-  loadExtensions: vi.fn(),
+vi.mock('../../config/extensions/storage.js', () => ({
   ExtensionStorage: {
     getUserExtensionsDir: vi.fn(),
   },
 }));
+vi.mock('../../config/extension-manager.js');
 vi.mock('@google/gemini-cli-core', () => ({
   createTransport: vi.fn(),
   MCPServerStatus: {
@@ -49,14 +50,18 @@ vi.mock('@modelcontextprotocol/sdk/client/index.js');
 const mockedGetUserExtensionsDir =
   ExtensionStorage.getUserExtensionsDir as Mock;
 const mockedLoadSettings = loadSettings as Mock;
-const mockedLoadExtensions = loadExtensions as Mock;
 const mockedCreateTransport = createTransport as Mock;
 const MockedClient = Client as Mock;
+const MockedExtensionManager = ExtensionManager as Mock;
 
 interface MockClient {
   connect: Mock;
   ping: Mock;
   close: Mock;
+}
+
+interface MockExtensionManager {
+  loadExtensions: Mock;
 }
 
 interface MockTransport {
@@ -66,6 +71,7 @@ interface MockTransport {
 describe('mcp list command', () => {
   let consoleSpy: MockInstance;
   let mockClient: MockClient;
+  let mockExtensionManager: MockExtensionManager;
   let mockTransport: MockTransport;
 
   beforeEach(() => {
@@ -79,10 +85,14 @@ describe('mcp list command', () => {
       ping: vi.fn(),
       close: vi.fn(),
     };
+    mockExtensionManager = {
+      loadExtensions: vi.fn(),
+    };
 
     MockedClient.mockImplementation(() => mockClient);
+    MockedExtensionManager.mockImplementation(() => mockExtensionManager);
     mockedCreateTransport.mockResolvedValue(mockTransport);
-    mockedLoadExtensions.mockReturnValue([]);
+    mockExtensionManager.loadExtensions.mockReturnValue([]);
     mockedGetUserExtensionsDir.mockReturnValue('/mocked/extensions/dir');
   });
 
@@ -159,7 +169,7 @@ describe('mcp list command', () => {
       },
     });
 
-    mockedLoadExtensions.mockReturnValue([
+    mockExtensionManager.loadExtensions.mockReturnValue([
       {
         name: 'test-extension',
         mcpServers: { 'extension-server': { command: '/ext/server' } },
