+++ packages/cli/src/config/extension.test.ts
@@ -4,36 +4,30 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { vi } from 'vitest';
+import { vi, type MockedFunction } from 'vitest';
 import * as fs from 'node:fs';
 import * as os from 'node:os';
 import * as path from 'node:path';
 import { createHash } from 'node:crypto';
 import {
-  EXTENSIONS_CONFIG_FILENAME,
-  ExtensionStorage,
-  INSTALL_METADATA_FILENAME,
-  INSTALL_WARNING_MESSAGE,
-  disableExtension,
-  enableExtension,
-  installOrUpdateExtension,
-  loadExtension,
-  loadExtensionConfig,
-  loadExtensions,
-  uninstallExtension,
-} from './extension.js';
-import {
-  GEMINI_DIR,
   type GeminiCLIExtension,
   ExtensionUninstallEvent,
   ExtensionDisableEvent,
   ExtensionEnableEvent,
 } from '@google/gemini-cli-core';
-import { SettingScope } from './settings.js';
+import { loadSettings, SettingScope } from './settings.js';
 import { isWorkspaceTrusted } from './trustedFolders.js';
 import { createExtension } from '../test-utils/createExtension.js';
 import { ExtensionEnablementManager } from './extensions/extensionEnablement.js';
 import { join } from 'node:path';
+import {
+  EXTENSIONS_CONFIG_FILENAME,
+  EXTENSIONS_DIRECTORY_NAME,
+  INSTALL_METADATA_FILENAME,
+} from './extensions/variables.js';
+import { ExtensionManager } from './extension-manager.js';
+import { ExtensionStorage } from './extensions/storage.js';
+import { INSTALL_WARNING_MESSAGE } from './extensions/consent.js';
 
 const mockGit = {
   clone: vi.fn(),
@@ -111,12 +105,12 @@ vi.mock('child_process', async (importOriginal) => {
   };
 });
 
-const EXTENSIONS_DIRECTORY_NAME = path.join(GEMINI_DIR, 'extensions');
-
 describe('extension tests', () => {
   let tempHomeDir: string;
   let tempWorkspaceDir: string;
   let userExtensionsDir: string;
+  let extensionManager: ExtensionManager;
+  let mockRequestConsent: MockedFunction<(consent: string) => Promise<boolean>>;
 
   beforeEach(() => {
     tempHomeDir = fs.mkdtempSync(
@@ -126,14 +120,21 @@ describe('extension tests', () => {
       path.join(tempHomeDir, 'gemini-cli-test-workspace-'),
     );
     userExtensionsDir = path.join(tempHomeDir, EXTENSIONS_DIRECTORY_NAME);
+    mockRequestConsent = vi.fn();
+    mockRequestConsent.mockResolvedValue(true);
     fs.mkdirSync(userExtensionsDir, { recursive: true });
-
     vi.mocked(os.homedir).mockReturnValue(tempHomeDir);
     vi.mocked(isWorkspaceTrusted).mockReturnValue({
       isTrusted: true,
       source: undefined,
     });
     vi.spyOn(process, 'cwd').mockReturnValue(tempWorkspaceDir);
+    extensionManager = new ExtensionManager({
+      workspaceDir: tempWorkspaceDir,
+      enabledExtensionOverrides: [],
+      requestConsent: mockRequestConsent,
+      loadedSettings: loadSettings(tempWorkspaceDir),
+    });
   });
 
   afterEach(() => {
@@ -153,7 +154,7 @@ describe('extension tests', () => {
         version: '1.0.0',
       });
 
-      const extensions = loadExtensions(new ExtensionEnablementManager());
+      const extensions = extensionManager.loadExtensions();
       expect(extensions).toHaveLength(1);
       expect(extensions[0].path).toBe(extensionDir);
       expect(extensions[0].name).toBe('test-extension');
@@ -172,7 +173,7 @@ describe('extension tests', () => {
         version: '2.0.0',
       });
 
-      const extensions = loadExtensions(new ExtensionEnablementManager());
+      const extensions = extensionManager.loadExtensions();
 
       expect(extensions).toHaveLength(2);
       const ext1 = extensions.find((e) => e.name === 'ext1');
@@ -192,7 +193,7 @@ describe('extension tests', () => {
         contextFileName: 'my-context-file.md',
       });
 
-      const extensions = loadExtensions(new ExtensionEnablementManager());
+      const extensions = extensionManager.loadExtensions();
 
       expect(extensions).toHaveLength(1);
       const ext1 = extensions.find((e) => e.name === 'ext1');
@@ -212,14 +213,11 @@ describe('extension tests', () => {
         name: 'enabled-extension',
         version: '2.0.0',
       });
-      const manager = new ExtensionEnablementManager();
-      disableExtension(
+      extensionManager.disableExtension(
         'disabled-extension',
         SettingScope.User,
-        manager,
-        tempWorkspaceDir,
       );
-      const extensions = loadExtensions(manager);
+      const extensions = extensionManager.loadExtensions();
       expect(extensions).toHaveLength(2);
       expect(extensions[0].name).toBe('disabled-extension');
       expect(extensions[0].isActive).toBe(false);
@@ -241,7 +239,7 @@ describe('extension tests', () => {
         },
       });
 
-      const extensions = loadExtensions(new ExtensionEnablementManager());
+      const extensions = extensionManager.loadExtensions();
       expect(extensions).toHaveLength(1);
       const expectedCwd = path.join(
         userExtensionsDir,
@@ -260,16 +258,13 @@ describe('extension tests', () => {
       });
       fs.writeFileSync(path.join(sourceExtDir, 'context.md'), 'linked context');
 
-      const extensionName = await installOrUpdateExtension(
-        {
-          source: sourceExtDir,
-          type: 'link',
-        },
-        async (_) => true,
-      );
+      const extensionName = await extensionManager.installOrUpdateExtension({
+        source: sourceExtDir,
+        type: 'link',
+      });
 
       expect(extensionName).toEqual('my-linked-extension');
-      const extensions = loadExtensions(new ExtensionEnablementManager());
+      const extensions = extensionManager.loadExtensions();
       expect(extensions).toHaveLength(1);
 
       const linkedExt = extensions[0];
@@ -318,7 +313,7 @@ describe('extension tests', () => {
         };
         fs.writeFileSync(configPath, JSON.stringify(extensionConfig));
 
-        const extensions = loadExtensions(new ExtensionEnablementManager());
+        const extensions = extensionManager.loadExtensions();
 
         expect(extensions).toHaveLength(1);
         const extension = extensions[0];
@@ -369,7 +364,7 @@ describe('extension tests', () => {
         JSON.stringify(extensionConfig),
       );
 
-      const extensions = loadExtensions(new ExtensionEnablementManager());
+      const extensions = extensionManager.loadExtensions();
 
       expect(extensions).toHaveLength(1);
       const extension = extensions[0];
@@ -397,7 +392,7 @@ describe('extension tests', () => {
       const badConfigPath = path.join(badExtDir, EXTENSIONS_CONFIG_FILENAME);
       fs.writeFileSync(badConfigPath, '{ "name": "bad-ext"'); // Malformed
 
-      const extensions = loadExtensions(new ExtensionEnablementManager());
+      const extensions = extensionManager.loadExtensions();
 
       expect(extensions).toHaveLength(1);
       expect(extensions[0].name).toBe('good-ext');
@@ -429,7 +424,7 @@ describe('extension tests', () => {
       const badConfigPath = path.join(badExtDir, EXTENSIONS_CONFIG_FILENAME);
       fs.writeFileSync(badConfigPath, JSON.stringify({ version: '1.0.0' }));
 
-      const extensions = loadExtensions(new ExtensionEnablementManager());
+      const extensions = extensionManager.loadExtensions();
 
       expect(extensions).toHaveLength(1);
       expect(extensions[0].name).toBe('good-ext');
@@ -457,7 +452,7 @@ describe('extension tests', () => {
         },
       });
 
-      const extensions = loadExtensions(new ExtensionEnablementManager());
+      const extensions = extensionManager.loadExtensions();
       expect(extensions).toHaveLength(1);
       expect(extensions[0].mcpServers?.['test-server'].trust).toBeUndefined();
     });
@@ -472,11 +467,7 @@ describe('extension tests', () => {
         version: '1.0.0',
       });
 
-      const extension = loadExtension({
-        extensionDir: badExtDir,
-        workspaceDir: tempWorkspaceDir,
-        extensionEnablementManager: new ExtensionEnablementManager(),
-      });
+      const extension = extensionManager.loadExtension(badExtDir);
 
       expect(extension).toBeNull();
       expect(consoleSpy).toHaveBeenCalledWith(
@@ -497,11 +488,7 @@ describe('extension tests', () => {
           },
         });
 
-        const extension = loadExtension({
-          extensionDir,
-          workspaceDir: tempWorkspaceDir,
-          extensionEnablementManager: new ExtensionEnablementManager(),
-        });
+        const extension = extensionManager.loadExtension(extensionDir);
 
         const expectedHash = createHash('sha256')
           .update('http://somehost.com/foo/bar')
@@ -520,11 +507,7 @@ describe('extension tests', () => {
           },
         });
 
-        const extension = loadExtension({
-          extensionDir,
-          workspaceDir: tempWorkspaceDir,
-          extensionEnablementManager: new ExtensionEnablementManager(),
-        });
+        const extension = extensionManager.loadExtension(extensionDir);
 
         const expectedHash = createHash('sha256')
           .update('https://github.com/foo/bar')
@@ -543,11 +526,7 @@ describe('extension tests', () => {
           },
         });
 
-        const extension = loadExtension({
-          extensionDir,
-          workspaceDir: tempWorkspaceDir,
-          extensionEnablementManager: new ExtensionEnablementManager(),
-        });
+        const extension = extensionManager.loadExtension(extensionDir);
 
         const expectedHash = createHash('sha256')
           .update('https://github.com/foo/bar')
@@ -566,11 +545,7 @@ describe('extension tests', () => {
           },
         });
 
-        const extension = loadExtension({
-          extensionDir,
-          workspaceDir: tempWorkspaceDir,
-          extensionEnablementManager: new ExtensionEnablementManager(),
-        });
+        const extension = extensionManager.loadExtension(extensionDir);
 
         const expectedHash = createHash('sha256')
           .update('https://github.com/foo/bar')
@@ -589,11 +564,7 @@ describe('extension tests', () => {
           },
         });
 
-        const extension = loadExtension({
-          extensionDir,
-          workspaceDir: tempWorkspaceDir,
-          extensionEnablementManager: new ExtensionEnablementManager(),
-        });
+        const extension = extensionManager.loadExtension(extensionDir);
 
         const expectedHash = createHash('sha256')
           .update('/some/path')
@@ -608,20 +579,14 @@ describe('extension tests', () => {
           name: 'link-ext-name',
           version: '1.0.0',
         });
-        const extensionName = await installOrUpdateExtension(
-          {
-            type: 'link',
-            source: actualExtensionDir,
-          },
-          async () => true,
-          tempWorkspaceDir,
-        );
-
-        const extension = loadExtension({
-          extensionDir: new ExtensionStorage(extensionName).getExtensionDir(),
-          workspaceDir: tempWorkspaceDir,
-          extensionEnablementManager: new ExtensionEnablementManager(),
+        const extensionName = await extensionManager.installOrUpdateExtension({
+          type: 'link',
+          source: actualExtensionDir,
         });
+
+        const extension = extensionManager.loadExtension(
+          new ExtensionStorage(extensionName).getExtensionDir(),
+        );
 
         const expectedHash = createHash('sha256')
           .update(actualExtensionDir)
@@ -636,11 +601,7 @@ describe('extension tests', () => {
           version: '1.0.0',
         });
 
-        const extension = loadExtension({
-          extensionDir,
-          workspaceDir: tempWorkspaceDir,
-          extensionEnablementManager: new ExtensionEnablementManager(),
-        });
+        const extension = extensionManager.loadExtension(extensionDir);
 
         const expectedHash = createHash('sha256')
           .update('no-meta-name')
@@ -660,10 +621,10 @@ describe('extension tests', () => {
       const targetExtDir = path.join(userExtensionsDir, 'my-local-extension');
       const metadataPath = path.join(targetExtDir, INSTALL_METADATA_FILENAME);
 
-      await installOrUpdateExtension(
-        { source: sourceExtDir, type: 'local' },
-        async (_) => true,
-      );
+      await extensionManager.installOrUpdateExtension({
+        source: sourceExtDir,
+        type: 'local',
+      });
 
       expect(fs.existsSync(targetExtDir)).toBe(true);
       expect(fs.existsSync(metadataPath)).toBe(true);
@@ -681,15 +642,15 @@ describe('extension tests', () => {
         name: 'my-local-extension',
         version: '1.0.0',
       });
-      await installOrUpdateExtension(
-        { source: sourceExtDir, type: 'local' },
-        async (_) => true,
-      );
+      await extensionManager.installOrUpdateExtension({
+        source: sourceExtDir,
+        type: 'local',
+      });
       await expect(
-        installOrUpdateExtension(
-          { source: sourceExtDir, type: 'local' },
-          async (_) => true,
-        ),
+        extensionManager.installOrUpdateExtension({
+          source: sourceExtDir,
+          type: 'local',
+        }),
       ).rejects.toThrow(
         'Extension "my-local-extension" is already installed. Please uninstall it first.',
       );
@@ -701,10 +662,10 @@ describe('extension tests', () => {
       const configPath = path.join(sourceExtDir, EXTENSIONS_CONFIG_FILENAME);
 
       await expect(
-        installOrUpdateExtension(
-          { source: sourceExtDir, type: 'local' },
-          async (_) => true,
-        ),
+        extensionManager.installOrUpdateExtension({
+          source: sourceExtDir,
+          type: 'local',
+        }),
       ).rejects.toThrow(`Configuration file not found at ${configPath}`);
 
       const targetExtDir = path.join(userExtensionsDir, 'bad-extension');
@@ -718,10 +679,10 @@ describe('extension tests', () => {
       fs.writeFileSync(configPath, '{ "name": "bad-json", "version": "1.0.0"'); // Malformed JSON
 
       await expect(
-        installOrUpdateExtension(
-          { source: sourceExtDir, type: 'local' },
-          async (_) => true,
-        ),
+        extensionManager.installOrUpdateExtension({
+          source: sourceExtDir,
+          type: 'local',
+        }),
       ).rejects.toThrow(
         new RegExp(
           `^Failed to load extension config from ${configPath.replace(
@@ -743,10 +704,10 @@ describe('extension tests', () => {
       fs.writeFileSync(configPath, JSON.stringify({ version: '1.0.0' }));
 
       await expect(
-        installOrUpdateExtension(
-          { source: sourceExtDir, type: 'local' },
-          async (_) => true,
-        ),
+        extensionManager.installOrUpdateExtension({
+          source: sourceExtDir,
+          type: 'local',
+        }),
       ).rejects.toThrow(
         `Invalid configuration in ${configPath}: missing "name"`,
       );
@@ -774,10 +735,10 @@ describe('extension tests', () => {
         type: 'github-release',
       });
 
-      await installOrUpdateExtension(
-        { source: gitUrl, type: 'git' },
-        async (_) => true,
-      );
+      await extensionManager.installOrUpdateExtension({
+        source: gitUrl,
+        type: 'git',
+      });
 
       expect(fs.existsSync(targetExtDir)).toBe(true);
       expect(fs.existsSync(metadataPath)).toBe(true);
@@ -798,10 +759,10 @@ describe('extension tests', () => {
       const metadataPath = path.join(targetExtDir, INSTALL_METADATA_FILENAME);
       const configPath = path.join(targetExtDir, EXTENSIONS_CONFIG_FILENAME);
 
-      await installOrUpdateExtension(
-        { source: sourceExtDir, type: 'link' },
-        async (_) => true,
-      );
+      await extensionManager.installOrUpdateExtension({
+        source: sourceExtDir,
+        type: 'link',
+      });
 
       expect(fs.existsSync(targetExtDir)).toBe(true);
       expect(fs.existsSync(metadataPath)).toBe(true);
@@ -828,20 +789,18 @@ describe('extension tests', () => {
             version: '1.1.0',
           });
           if (isUpdate) {
-            await installOrUpdateExtension(
-              { source: sourceExtDir, type: 'local' },
-              async (_) => true,
-            );
+            await extensionManager.installOrUpdateExtension({
+              source: sourceExtDir,
+              type: 'local',
+            });
           }
           // Clears out any calls to mocks from the above function calls.
           vi.clearAllMocks();
         });
 
         it(`should log an ${isUpdate ? 'update' : 'install'} event to clearcut on success`, async () => {
-          await installOrUpdateExtension(
+          await extensionManager.installOrUpdateExtension(
             { source: sourceExtDir, type: 'local' },
-            async (_) => true,
-            undefined,
             isUpdate
               ? {
                   name: 'my-local-extension',
@@ -863,10 +822,8 @@ describe('extension tests', () => {
           const enablementManager = new ExtensionEnablementManager();
           enablementManager.enable('my-local-extension', true, '/some/scope');
 
-          await installOrUpdateExtension(
+          await extensionManager.installOrUpdateExtension(
             { source: sourceExtDir, type: 'local' },
-            async (_) => true,
-            undefined,
             isUpdate
               ? {
                   name: 'my-local-extension',
@@ -904,14 +861,11 @@ describe('extension tests', () => {
         },
       });
 
-      const mockRequestConsent = vi.fn();
-      mockRequestConsent.mockResolvedValue(true);
-
       await expect(
-        installOrUpdateExtension(
-          { source: sourceExtDir, type: 'local' },
-          mockRequestConsent,
-        ),
+        extensionManager.installOrUpdateExtension({
+          source: sourceExtDir,
+          type: 'local',
+        }),
       ).resolves.toBe('my-local-extension');
 
       expect(mockRequestConsent).toHaveBeenCalledWith(
@@ -937,10 +891,10 @@ This extension will run the following MCP servers:
       });
 
       await expect(
-        installOrUpdateExtension(
-          { source: sourceExtDir, type: 'local' },
-          async () => true,
-        ),
+        extensionManager.installOrUpdateExtension({
+          source: sourceExtDir,
+          type: 'local',
+        }),
       ).resolves.toBe('my-local-extension');
     });
 
@@ -956,12 +910,12 @@ This extension will run the following MCP servers:
           },
         },
       });
-
+      mockRequestConsent.mockResolvedValue(false);
       await expect(
-        installOrUpdateExtension(
-          { source: sourceExtDir, type: 'local' },
-          async () => false,
-        ),
+        extensionManager.installOrUpdateExtension({
+          source: sourceExtDir,
+          type: 'local',
+        }),
       ).rejects.toThrow('Installation cancelled for "my-local-extension".');
     });
 
@@ -974,14 +928,11 @@ This extension will run the following MCP servers:
       const targetExtDir = path.join(userExtensionsDir, 'my-local-extension');
       const metadataPath = path.join(targetExtDir, INSTALL_METADATA_FILENAME);
 
-      await installOrUpdateExtension(
-        {
-          source: sourceExtDir,
-          type: 'local',
-          autoUpdate: true,
-        },
-        async (_) => true,
-      );
+      await extensionManager.installOrUpdateExtension({
+        source: sourceExtDir,
+        type: 'local',
+        autoUpdate: true,
+      });
 
       expect(fs.existsSync(targetExtDir)).toBe(true);
       expect(fs.existsSync(metadataPath)).toBe(true);
@@ -1007,29 +958,24 @@ This extension will run the following MCP servers:
         },
       });
 
-      const mockRequestConsent = vi.fn();
-      // Install it and force consent first.
-      await installOrUpdateExtension(
-        { source: sourceExtDir, type: 'local' },
-        async () => true,
-      );
+      // Install it with hard coded consent first.
+      await extensionManager.installOrUpdateExtension({
+        source: sourceExtDir,
+        type: 'local',
+      });
+      expect(mockRequestConsent).toHaveBeenCalledOnce();
 
       // Now update it without changing anything.
       await expect(
-        installOrUpdateExtension(
+        extensionManager.installOrUpdateExtension(
           { source: sourceExtDir, type: 'local' },
-          mockRequestConsent,
-          process.cwd(),
           // Provide its own existing config as the previous config.
-          await loadExtensionConfig({
-            extensionDir: sourceExtDir,
-            workspaceDir: process.cwd(),
-            extensionEnablementManager: new ExtensionEnablementManager(),
-          }),
+          await extensionManager.loadExtensionConfig(sourceExtDir),
         ),
       ).resolves.toBe('my-local-extension');
 
-      expect(mockRequestConsent).not.toHaveBeenCalled();
+      // Still only called once
+      expect(mockRequestConsent).toHaveBeenCalledOnce();
     });
 
     it('should throw an error for invalid extension names', async () => {
@@ -1040,10 +986,10 @@ This extension will run the following MCP servers:
       });
 
       await expect(
-        installOrUpdateExtension(
-          { source: sourceExtDir, type: 'local' },
-          async (_) => true,
-        ),
+        extensionManager.installOrUpdateExtension({
+          source: sourceExtDir,
+          type: 'local',
+        }),
       ).rejects.toThrow('Invalid extension name: "bad_name"');
     });
 
@@ -1088,10 +1034,10 @@ This extension will run the following MCP servers:
           join(tempDir, extensionName),
         );
 
-        await installOrUpdateExtension(
-          { source: gitUrl, type: 'github-release' },
-          async () => true,
-        );
+        await extensionManager.installOrUpdateExtension({
+          source: gitUrl,
+          type: 'github-release',
+        });
 
         expect(fs.existsSync(targetExtDir)).toBe(true);
         const metadataPath = path.join(targetExtDir, INSTALL_METADATA_FILENAME);
@@ -1111,17 +1057,15 @@ This extension will run the following MCP servers:
           errorMessage: 'download failed',
           type: 'github-release',
         });
-        const requestConsent = vi.fn().mockResolvedValue(true);
 
-        await installOrUpdateExtension(
+        await extensionManager.installOrUpdateExtension(
           { source: gitUrl, type: 'github-release' }, // Use github-release to force consent
-          requestConsent,
         );
 
         // It gets called once to ask for a git clone, and once to consent to
         // the actual extension features.
-        expect(requestConsent).toHaveBeenCalledTimes(2);
-        expect(requestConsent).toHaveBeenCalledWith(
+        expect(mockRequestConsent).toHaveBeenCalledTimes(2);
+        expect(mockRequestConsent).toHaveBeenCalledWith(
           expect.stringContaining(
             'Would you like to attempt to install via "git clone" instead?',
           ),
@@ -1142,18 +1086,18 @@ This extension will run the following MCP servers:
           errorMessage: 'download failed',
           type: 'github-release',
         });
-        const requestConsent = vi.fn().mockResolvedValue(false);
+        mockRequestConsent.mockResolvedValue(false);
 
         await expect(
-          installOrUpdateExtension(
-            { source: gitUrl, type: 'github-release' },
-            requestConsent,
-          ),
+          extensionManager.installOrUpdateExtension({
+            source: gitUrl,
+            type: 'github-release',
+          }),
         ).rejects.toThrow(
           `Failed to install extension ${gitUrl}: download failed`,
         );
 
-        expect(requestConsent).toHaveBeenCalledExactlyOnceWith(
+        expect(mockRequestConsent).toHaveBeenCalledExactlyOnceWith(
           expect.stringContaining(
             'Would you like to attempt to install via "git clone" instead?',
           ),
@@ -1167,16 +1111,15 @@ This extension will run the following MCP servers:
           failureReason: 'no release data',
           type: 'github-release',
         });
-        const requestConsent = vi.fn().mockResolvedValue(true);
 
-        await installOrUpdateExtension(
-          { source: gitUrl, type: 'git' },
-          requestConsent,
-        );
+        await extensionManager.installOrUpdateExtension({
+          source: gitUrl,
+          type: 'git',
+        });
 
         // We should not see the request to use git clone, this is a repo that
         // has no github releases so it is the only install method.
-        expect(requestConsent).toHaveBeenCalledExactlyOnceWith(
+        expect(mockRequestConsent).toHaveBeenCalledExactlyOnceWith(
           expect.stringContaining(
             'Installing extension "gemini-test-extension"',
           ),
@@ -1198,14 +1141,12 @@ This extension will run the following MCP servers:
           errorMessage: 'No release data found',
           type: 'github-release',
         });
-        const requestConsent = vi.fn().mockResolvedValue(true);
 
-        await installOrUpdateExtension(
+        await extensionManager.installOrUpdateExtension(
           { source: gitUrl, type: 'github-release' }, // Note the type
-          requestConsent,
         );
 
-        expect(requestConsent).toHaveBeenCalledWith(
+        expect(mockRequestConsent).toHaveBeenCalledWith(
           expect.stringContaining(
             'Would you like to attempt to install via "git clone" instead?',
           ),
@@ -1223,7 +1164,7 @@ This extension will run the following MCP servers:
         version: '1.0.0',
       });
 
-      await uninstallExtension('my-local-extension', false);
+      await extensionManager.uninstallExtension('my-local-extension', false);
 
       expect(fs.existsSync(sourceExtDir)).toBe(false);
     });
@@ -1240,16 +1181,16 @@ This extension will run the following MCP servers:
         version: '1.0.0',
       });
 
-      await uninstallExtension('my-local-extension', false);
+      await extensionManager.uninstallExtension('my-local-extension', false);
 
       expect(fs.existsSync(sourceExtDir)).toBe(false);
-      expect(loadExtensions(new ExtensionEnablementManager())).toHaveLength(1);
+      expect(extensionManager.loadExtensions()).toHaveLength(1);
       expect(fs.existsSync(otherExtDir)).toBe(true);
     });
 
     it('should throw an error if the extension does not exist', async () => {
       await expect(
-        uninstallExtension('nonexistent-extension', false),
+        extensionManager.uninstallExtension('nonexistent-extension', false),
       ).rejects.toThrow('Extension not found.');
     });
 
@@ -1261,7 +1202,10 @@ This extension will run the following MCP servers:
           version: '1.0.0',
         });
 
-        await uninstallExtension('my-local-extension', isUpdate);
+        await extensionManager.uninstallExtension(
+          'my-local-extension',
+          isUpdate,
+        );
 
         if (isUpdate) {
           expect(mockLogExtensionUninstall).not.toHaveBeenCalled();
@@ -1284,7 +1228,7 @@ This extension will run the following MCP servers:
         const enablementManager = new ExtensionEnablementManager();
         enablementManager.enable('test-extension', true, '/some/scope');
 
-        await uninstallExtension('test-extension', isUpdate);
+        await extensionManager.uninstallExtension('test-extension', isUpdate);
 
         const config = enablementManager.readConfig()['test-extension'];
         if (isUpdate) {
@@ -1308,7 +1252,7 @@ This extension will run the following MCP servers:
         },
       });
 
-      await uninstallExtension(gitUrl, false);
+      await extensionManager.uninstallExtension(gitUrl, false);
 
       expect(fs.existsSync(sourceExtDir)).toBe(false);
       expect(mockLogExtensionUninstall).toHaveBeenCalled();
@@ -1327,7 +1271,7 @@ This extension will run the following MCP servers:
       });
 
       await expect(
-        uninstallExtension(
+        extensionManager.uninstallExtension(
           'https://github.com/google/no-metadata-extension',
           false,
         ),
@@ -1343,11 +1287,7 @@ This extension will run the following MCP servers:
         version: '1.0.0',
       });
 
-      disableExtension(
-        'my-extension',
-        SettingScope.User,
-        new ExtensionEnablementManager(),
-      );
+      extensionManager.disableExtension('my-extension', SettingScope.User);
       expect(
         isEnabled({
           name: 'my-extension',
@@ -1363,12 +1303,7 @@ This extension will run the following MCP servers:
         version: '1.0.0',
       });
 
-      disableExtension(
-        'my-extension',
-        SettingScope.Workspace,
-        new ExtensionEnablementManager(),
-        tempWorkspaceDir,
-      );
+      extensionManager.disableExtension('my-extension', SettingScope.Workspace);
       expect(
         isEnabled({
           name: 'my-extension',
@@ -1390,16 +1325,8 @@ This extension will run the following MCP servers:
         version: '1.0.0',
       });
 
-      disableExtension(
-        'my-extension',
-        SettingScope.User,
-        new ExtensionEnablementManager(),
-      );
-      disableExtension(
-        'my-extension',
-        SettingScope.User,
-        new ExtensionEnablementManager(),
-      );
+      extensionManager.disableExtension('my-extension', SettingScope.User);
+      extensionManager.disableExtension('my-extension', SettingScope.User);
       expect(
         isEnabled({
           name: 'my-extension',
@@ -1410,11 +1337,7 @@ This extension will run the following MCP servers:
 
     it('should throw an error if you request system scope', () => {
       expect(() =>
-        disableExtension(
-          'my-extension',
-          SettingScope.System,
-          new ExtensionEnablementManager(),
-        ),
+        extensionManager.disableExtension('my-extension', SettingScope.System),
       ).toThrow('System and SystemDefaults scopes are not supported.');
     });
 
@@ -1425,11 +1348,7 @@ This extension will run the following MCP servers:
         version: '1.0.0',
       });
 
-      disableExtension(
-        'ext1',
-        SettingScope.Workspace,
-        new ExtensionEnablementManager(),
-      );
+      extensionManager.disableExtension('ext1', SettingScope.Workspace);
 
       expect(mockLogExtensionDisable).toHaveBeenCalled();
       expect(ExtensionDisableEvent).toHaveBeenCalledWith(
@@ -1445,8 +1364,7 @@ This extension will run the following MCP servers:
     });
 
     const getActiveExtensions = (): GeminiCLIExtension[] => {
-      const manager = new ExtensionEnablementManager();
-      const extensions = loadExtensions(manager);
+      const extensions = extensionManager.loadExtensions();
       return extensions.filter((e) => e.isActive);
     };
 
@@ -1456,12 +1374,11 @@ This extension will run the following MCP servers:
         name: 'ext1',
         version: '1.0.0',
       });
-      const extensionEnablementManager = new ExtensionEnablementManager();
-      disableExtension('ext1', SettingScope.User, extensionEnablementManager);
+      extensionManager.disableExtension('ext1', SettingScope.User);
       let activeExtensions = getActiveExtensions();
       expect(activeExtensions).toHaveLength(0);
 
-      enableExtension('ext1', SettingScope.User, extensionEnablementManager);
+      extensionManager.enableExtension('ext1', SettingScope.User);
       activeExtensions = getActiveExtensions();
       expect(activeExtensions).toHaveLength(1);
       expect(activeExtensions[0].name).toBe('ext1');
@@ -1473,20 +1390,11 @@ This extension will run the following MCP servers:
         name: 'ext1',
         version: '1.0.0',
       });
-      const extensionEnablementManager = new ExtensionEnablementManager();
-      disableExtension(
-        'ext1',
-        SettingScope.Workspace,
-        extensionEnablementManager,
-      );
+      extensionManager.disableExtension('ext1', SettingScope.Workspace);
       let activeExtensions = getActiveExtensions();
       expect(activeExtensions).toHaveLength(0);
 
-      enableExtension(
-        'ext1',
-        SettingScope.Workspace,
-        extensionEnablementManager,
-      );
+      extensionManager.enableExtension('ext1', SettingScope.Workspace);
       activeExtensions = getActiveExtensions();
       expect(activeExtensions).toHaveLength(1);
       expect(activeExtensions[0].name).toBe('ext1');
@@ -1498,17 +1406,8 @@ This extension will run the following MCP servers:
         name: 'ext1',
         version: '1.0.0',
       });
-      const extensionEnablementManager = new ExtensionEnablementManager();
-      disableExtension(
-        'ext1',
-        SettingScope.Workspace,
-        extensionEnablementManager,
-      );
-      enableExtension(
-        'ext1',
-        SettingScope.Workspace,
-        extensionEnablementManager,
-      );
+      extensionManager.disableExtension('ext1', SettingScope.Workspace);
+      extensionManager.enableExtension('ext1', SettingScope.Workspace);
 
       expect(mockLogExtensionEnable).toHaveBeenCalled();
       expect(ExtensionEnableEvent).toHaveBeenCalledWith(
