+++ packages/cli/src/config/extensions/update.ts
@@ -9,19 +9,13 @@ import {
   ExtensionUpdateState,
   type ExtensionUpdateStatus,
 } from '../../ui/state/extensions.js';
-import {
-  copyExtension,
-  installOrUpdateExtension,
-  loadExtension,
-  loadInstallMetadata,
-  ExtensionStorage,
-  loadExtensionConfig,
-} from '../extension.js';
+import { loadInstallMetadata } from '../extension.js';
 import { checkForExtensionUpdate } from './github.js';
 import { debugLogger, type GeminiCLIExtension } from '@google/gemini-cli-core';
 import * as fs from 'node:fs';
 import { getErrorMessage } from '../../utils/errors.js';
-import { type ExtensionEnablementManager } from './extensionEnablement.js';
+import { copyExtension, type ExtensionManager } from '../extension-manager.js';
+import { ExtensionStorage } from './storage.js';
 
 export interface ExtensionUpdateInfo {
   name: string;
@@ -31,9 +25,7 @@ export interface ExtensionUpdateInfo {
 
 export async function updateExtension(
   extension: GeminiCLIExtension,
-  extensionEnablementManager: ExtensionEnablementManager,
-  cwd: string = process.cwd(),
-  requestConsent: (consent: string) => Promise<boolean>,
+  extensionManager: ExtensionManager,
   currentState: ExtensionUpdateState,
   dispatchExtensionStateUpdate: (action: ExtensionUpdateAction) => void,
 ): Promise<ExtensionUpdateInfo | undefined> {
@@ -66,24 +58,18 @@ export async function updateExtension(
 
   const tempDir = await ExtensionStorage.createTmpDir();
   try {
-    const previousExtensionConfig = await loadExtensionConfig({
-      extensionDir: extension.path,
-      workspaceDir: cwd,
-      extensionEnablementManager,
-    });
-    await installOrUpdateExtension(
+    const previousExtensionConfig = await extensionManager.loadExtensionConfig(
+      extension.path,
+    );
+    await extensionManager.installOrUpdateExtension(
       installMetadata,
-      requestConsent,
-      cwd,
       previousExtensionConfig,
     );
 
     const updatedExtensionStorage = new ExtensionStorage(extension.name);
-    const updatedExtension = loadExtension({
-      extensionDir: updatedExtensionStorage.getExtensionDir(),
-      workspaceDir: cwd,
-      extensionEnablementManager,
-    });
+    const updatedExtension = extensionManager.loadExtension(
+      updatedExtensionStorage.getExtensionDir(),
+    );
     if (!updatedExtension) {
       dispatchExtensionStateUpdate({
         type: 'SET_STATE',
@@ -120,11 +106,9 @@ export async function updateExtension(
 }
 
 export async function updateAllUpdatableExtensions(
-  cwd: string = process.cwd(),
-  requestConsent: (consent: string) => Promise<boolean>,
   extensions: GeminiCLIExtension[],
   extensionsState: Map<string, ExtensionUpdateStatus>,
-  extensionEnablementManager: ExtensionEnablementManager,
+  extensionManager: ExtensionManager,
   dispatch: (action: ExtensionUpdateAction) => void,
 ): Promise<ExtensionUpdateInfo[]> {
   return (
@@ -138,9 +122,7 @@ export async function updateAllUpdatableExtensions(
         .map((extension) =>
           updateExtension(
             extension,
-            extensionEnablementManager,
-            cwd,
-            requestConsent,
+            extensionManager,
             extensionsState.get(extension.name)!.status,
             dispatch,
           ),
@@ -156,9 +138,8 @@ export interface ExtensionUpdateCheckResult {
 
 export async function checkForAllExtensionUpdates(
   extensions: GeminiCLIExtension[],
-  extensionEnablementManager: ExtensionEnablementManager,
+  extensionManager: ExtensionManager,
   dispatch: (action: ExtensionUpdateAction) => void,
-  cwd: string = process.cwd(),
 ): Promise<void> {
   dispatch({ type: 'BATCH_CHECK_START' });
   const promises: Array<Promise<void>> = [];
@@ -181,12 +162,11 @@ export async function checkForAllExtensionUpdates(
       },
     });
     promises.push(
-      checkForExtensionUpdate(extension, extensionEnablementManager, cwd).then(
-        (state) =>
-          dispatch({
-            type: 'SET_STATE',
-            payload: { name: extension.name, state },
-          }),
+      checkForExtensionUpdate(extension, extensionManager).then((state) =>
+        dispatch({
+          type: 'SET_STATE',
+          payload: { name: extension.name, state },
+        }),
       ),
     );
   }
