+++ packages/cli/src/config/extensions/variableSchema.ts
@@ -4,8 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import type { ExtensionEnablementManager } from './extensionEnablement.js';
-
 export interface VariableDefinition {
   type: 'string';
   description: string;
@@ -15,12 +13,6 @@ export interface VariableDefinition {
 
 export interface VariableSchema {
   [key: string]: VariableDefinition;
-}
-
-export interface LoadExtensionContext {
-  extensionDir: string;
-  workspaceDir: string;
-  extensionEnablementManager: ExtensionEnablementManager;
 }
 
 const PATH_SEPARATOR_DEFINITION = {
