+++ packages/cli/src/ui/hooks/useExtensionUpdates.ts
@@ -18,12 +18,9 @@ import {
   checkForAllExtensionUpdates,
   updateExtension,
 } from '../../config/extensions/update.js';
-import {
-  requestConsentInteractive,
-  type ExtensionUpdateInfo,
-} from '../../config/extension.js';
+import { type ExtensionUpdateInfo } from '../../config/extension.js';
 import { checkExhaustive } from '../../utils/checks.js';
-import type { ExtensionEnablementManager } from '../../config/extensions/extensionEnablement.js';
+import type { ExtensionManager } from '../../config/extension-manager.js';
 
 type ConfirmationRequestWrapper = {
   prompt: React.ReactNode;
@@ -48,16 +45,7 @@ function confirmationRequestsReducer(
   }
 }
 
-export const useExtensionUpdates = (
-  extensions: GeminiCLIExtension[],
-  extensionEnablementManager: ExtensionEnablementManager,
-  addItem: UseHistoryManagerReturn['addItem'],
-  cwd: string,
-) => {
-  const [extensionsUpdateState, dispatchExtensionStateUpdate] = useReducer(
-    extensionUpdatesReducer,
-    initialExtensionUpdatesState,
-  );
+export const useConfirmUpdateRequests = () => {
   const [
     confirmUpdateExtensionRequests,
     dispatchConfirmUpdateExtensionRequests,
@@ -82,6 +70,22 @@ export const useExtensionUpdates = (
     },
     [dispatchConfirmUpdateExtensionRequests],
   );
+  return {
+    addConfirmUpdateExtensionRequest,
+    confirmUpdateExtensionRequests,
+    dispatchConfirmUpdateExtensionRequests,
+  };
+};
+
+export const useExtensionUpdates = (
+  extensions: GeminiCLIExtension[],
+  extensionManager: ExtensionManager,
+  addItem: UseHistoryManagerReturn['addItem'],
+) => {
+  const [extensionsUpdateState, dispatchExtensionStateUpdate] = useReducer(
+    extensionUpdatesReducer,
+    initialExtensionUpdatesState,
+  );
 
   useEffect(() => {
     const extensionsToCheck = extensions.filter((extension) => {
@@ -95,15 +99,13 @@ export const useExtensionUpdates = (
     if (extensionsToCheck.length === 0) return;
     checkForAllExtensionUpdates(
       extensionsToCheck,
-      extensionEnablementManager,
+      extensionManager,
       dispatchExtensionStateUpdate,
-      cwd,
     );
   }, [
     extensions,
-    extensionEnablementManager,
+    extensionManager,
     extensionsUpdateState.extensionStatuses,
-    cwd,
     dispatchExtensionStateUpdate,
   ]);
 
@@ -158,13 +160,7 @@ export const useExtensionUpdates = (
       } else {
         const updatePromise = updateExtension(
           extension,
-          extensionEnablementManager,
-          cwd,
-          (description) =>
-            requestConsentInteractive(
-              description,
-              addConfirmUpdateExtensionRequest,
-            ),
+          extensionManager,
           currentState.status,
           dispatchExtensionStateUpdate,
         );
@@ -213,14 +209,7 @@ export const useExtensionUpdates = (
         });
       });
     }
-  }, [
-    extensions,
-    extensionEnablementManager,
-    extensionsUpdateState,
-    addConfirmUpdateExtensionRequest,
-    addItem,
-    cwd,
-  ]);
+  }, [extensions, extensionManager, extensionsUpdateState, addItem]);
 
   const extensionsUpdateStateComputed = useMemo(() => {
     const result = new Map<string, ExtensionUpdateState>();
@@ -237,7 +226,5 @@ export const useExtensionUpdates = (
     extensionsUpdateState: extensionsUpdateStateComputed,
     extensionsUpdateStateInternal: extensionsUpdateState.extensionStatuses,
     dispatchExtensionStateUpdate,
-    confirmUpdateExtensionRequests,
-    addConfirmUpdateExtensionRequest,
   };
 };
